+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [5.1.9] - 2024-01-16
+### Fixed
+- Fixed content update check in ContentAttributesPanel [#3828](https://github.com/rokwire/illinois-app/issues/3828).
+- Fixed Event2 location detail display [#3842](https://github.com/rokwire/illinois-app/issues/3842) [#3843](https://github.com/rokwire/illinois-app/issues/3843).
+
 ## [5.1.8] - 2024-01-15
 ### Changed
 - Updated some display strings in Event2QRCodePanel [#3838](https://github.com/rokwire/illinois-app/issues/3838).